chore: upgrade Next.js to v16
