+++ next.config.ts
@@ -1,10 +1,7 @@
-import type { NextConfig } from 'next';
+import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  experimental: {
-    ppr: true,
-    clientSegmentCache: true
-  }
+  cacheComponents: true,
 };
 
 export default nextConfig;
